log: Move redaction related code to separate file.
