+++ lib/log/redact_test.go
